fix(delivery-service): BigDecimal for delivery_partners.rating (validate); ECR Public base images for CodeBuild

+++ services/delivery-service/src/main/java/com/warmpawz/delivery/service/serviceimpl/DeliveryTrackingServiceImpl.java
@@ -384,7 +384,7 @@ public class DeliveryTrackingServiceImpl implements DeliveryTrackingService {
 		p.setVehicleNumber(vehicleNumber);
 		p.setCurrentLocationJson("{\"lat\":12.9716,\"lng\":77.5946}");
 		p.setStatus("available");
-		p.setRating(5.0);
+		p.setRating(BigDecimal.valueOf(5));
 		p.setTotalDeliveries(0);
 		p.setActive(true);
 		try {
